Move Curruent and Voltage sense to Io
